Macro rlg::macro_print_log

macro_rules! macro_print_log {
    ($log:expr) => { ... };
}

This macro prints a log entry to the standard output (stdout). It is useful for debugging or simple logging to the console.

Parameters

log: The log entry to be printed.

Example

use rlg::{macro_print_log,macro_info_log};
let log = macro_info_log!("2022-01-01", "app", "message");
macro_print_log!(log);

Usage: macro_print_log!(log); Print log to stdout