Macro rlg::macro_log_to_file

macro_rules! macro_log_to_file {
    ($log:expr) => { ... };
}

This macro asynchronously logs a message to a file. It returns the result of the logging operation, which could be used to check the success or failure of the logging action.

Parameters

log: The log entry to be saved to a file.

Example

use rlg::{macro_log,macro_log_to_file,macro_info_log};
use rlg::log_format::LogFormat;
use rlg::log_level::LogLevel;
let log = macro_info_log!("2022-01-01", "app", "message");
async {
let result = macro_log_to_file!(log);
};

Usage: let result = macro_log_to_file!(log); Async log message to file